requirements-backend.txt
# Backend API Requirements

fastapi==0.110.0
uvicorn[standard]==0.28.0
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.1
redis==5.1.1
pydantic-settings==2.1.0
